fix: feedback corrections
close #1

+++ src/n1exercici1/Product.java
@@ -1,10 +1,12 @@
 package n1exercici1;
 
+import java.math.BigDecimal;
+
 public class Product {
     private String name;
-    private double price;
+    private BigDecimal price;
 
-    public Product(String name, double price){
+    public Product(String name, BigDecimal price){
         this.name = name;
         this.price = price;
     }
@@ -12,7 +14,7 @@ public class Product {
     public String getName() {
         return name;
     }
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
@@ -20,7 +22,7 @@ public class Product {
         this.name = name;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
 }
